requirements.txt
langdetect==1.0.9
pyspark==3.2.1
wget==3.2
psycopg2==2.9.3
tqdm==4.64.0
config==0.5.1
PyYAML==6.0
pgvector==0.1.5
SQLAlchemy==1.4.35
fastapi==0.75.2
pydantic==1.9.0